Merge configurations when bundles are provided multiple times
When two extensions provide the same bundle to the system and both
provide a config for `loadAfter` and `replace`, the values must get
merged to a common array.
When at least one of each sets `loadInProduction` or
`loadInDevelopment`, the result must also get loaded then and not only
the one coming arbitrary "last".

Variance in config types is explicitely not handled so far and marked as
FIXME for the moment as I have no clue how to handle these and the use
case seems too narrow to handle them explicitely.