Fix issue that loading order permutates per system (see #27)
Description
-----------

As discussed on Mumble on 2020-07-02, we want reproducible builds (as everyone else).
However, Contao creates a `bundles.map` which can vary per host, directory, etc. despite using the very same `composer.lock` for input.
This fixes that behaviour by sorting the loading order keys and therefore have an deterministic output of the resolved dependencies.

Commits
-------

91d11735 Fix issue that loading order permutates per system

As discussed on Mumble on 2020-07-02, we want reproducible builds (as everyone else).
However, Contao creates a `bundles.map` which can vary per host, directory, etc. despite using the very same `composer.lock` for input.
This fixes that behaviour by sorting the loading order keys and therefore have an deterministic output of the resolved dependencies.
e510734b Merge configurations when bundles are provided multiple times

When two extensions provide the same bundle to the system and both
provide a config for `loadAfter` and `replace`, the values must get
merged to a common array.
When at least one of each sets `loadInProduction` or
`loadInDevelopment`, the result must also get loaded then and not only
the one coming arbitrary "last".

Variance in config types is explicitely not handled so far and marked as
FIXME for the moment as I have no clue how to handle these and the use
case seems too narrow to handle them explicitely.
bdc7339c Handle `setReplace` and  `setLoadAfter` sorted

To have predictable output, the results must be sorted, for
functionality the order does not matter so this is fine.
d9788e12 Fix coding standard issues
dfb04e78 Fix for assertEquals => assertSame in test

Now we have `assertSame` back at the expense of needing a loop.
However, the result is now guaranteed to exactly match the expected
output in order and contents.
8d8207b6 Yet another coding standard issue
172bedff Runtests five times with mt_srand(0) til mt_srand(4)

We now run the tests five times with different random seeds to ensure
the random sorting is predictable as well.
This is a compromise between having "random" input values and
predictable test output.
587f9458 Drop FIXME

After discussion we decided to keep the exception for the moment and
revisit this as soon as some one encounteres a problem with it.
5561c937 Add conflict with Contao <4.9.4

We now conflict with all previous 4.9 manager-bundle versions, as these
do not contain the fixes from contao/contao#1903 which are needed to not
break things for existing installations.
96fc04c3 Replace `ksort` with `md5` comparison
4eb0b15c CI