Add provider repositories on require command (see #54)
Description
-----------

Our Composer plugin dynamically registers repository URLs for packages in the artifact repository `/contao-manager/packages`. Currently, it only adds repositories for packages that are required (= are listed in the root `composer.json`). 

This pull request extends the capability to also register repository URLs in the `composer require` command. This fixes the problem that you cannot require a `contao-provider` and a package provided by that provider at the same time.

Commits
-------

fb3ed312 Add provider repositories on require command
0fbf36c4 Use the Composer VersionParser
1a8d6f5a Ping
86a46ea0 Pong
2a5356ee Raise Composer dependency to 1.7 for pre-command-run event